transfer function based on the derivative of the bank
