+++ driver/ahrs.h
@@ -1,0 +1,8 @@
+#ifndef AHRS_H
+#define AHRS_H
+
+#include "state.h"
+
+void refresh_ahrs(struct state_t *state);
+
+#endif
