+++ driver/automated.h
@@ -1,0 +1,6 @@
+#ifndef AUTOMATED_H
+#define AUTOMATED_H
+
+int automated_loop(struct state_t *state);
+
+#endif
