+++ driver/setup.h
@@ -1,0 +1,9 @@
+#ifndef SETUP_H
+#define SETUP_H
+
+#include "setup.h"
+
+int setup_state(struct state_t *state);
+void cleanup_state(struct state_t *state);
+
+#endif
