+++ driver/utils.h
@@ -1,0 +1,6 @@
+#ifndef UTILS_H
+#define UTILS_H
+
+int kbhit();
+
+#endif
